test 4

+++ src/drivers/src/i2cdev_f405.c
@@ -106,7 +106,7 @@ bool i2cdevReadReg16(I2C_Dev *dev, uint8_t devAddress, uint16_t memAddress,
 {
   I2cMessage message;
   DEBUG_PRINT("Creating message address... Device 0x%x\tMemory 0x%x\tLength 0x%x\tData 0x%x",
-              devAddress, memAddress, len, data);
+              devAddress, memAddress, len, (uint8_t)data);
   i2cdrvCreateMessageIntAddr(&message, devAddress, true, memAddress,
                           i2cRead, len, data);
   DEBUG_PRINT("Done!\n");
